updated the merge_df function to merge the custom aggregated tables provided by Max and Keji.
